Added More bones to closestBone Aimbot. New Bone Selection Popup. Put all the Bone ID's into enums for each player model. Util functions for getting Player Model Types. Config/saving for new bones. Fixed an issue where the aimbot would follow through walls constantly with engageLock on.(Now only does so for 250ms)

+++ src/Hacks/dlights.cpp
@@ -36,7 +36,7 @@ void Dlights::Paint()
 		bool bIsVisible = false;
 		if (Settings::ESP::Filters::visibilityCheck || Settings::ESP::Filters::legit)
 		{
-			bIsVisible = Entity::IsVisible(player, Bone::BONE_HEAD, 180.f, Settings::ESP::Filters::smokeCheck);
+			bIsVisible = Entity::IsVisible(player, (int)Bone::BONE_HEAD, 180.f, Settings::ESP::Filters::smokeCheck);
 			if (!bIsVisible && Settings::ESP::Filters::legit)
 				continue;
 		}
